Create Regression
Regression is used to predict a (real) number. In the example in the cell below we have 12 data points in the variable data. The variables prediction and alternative are the predictions by two linear regression models that fit the data. In the plot, you can see both predictions and the original data.