Create Nodes(Decision Tree)
The cell below defines functions we can use to construct nodes (represented as dictionaries) in a decision tree. The nodes store references to its children (left and right) and to the question (feature value) to ask about at this node, as well as the class we'll predict at this node.